+++ Engine/cs_par_orc_semantic_analysis.c
@@ -27,13 +27,11 @@
 #include "csoundCore.h"
 #include "csound_orc.h"
 #include "tok.h"
-
 #include "cs_par_base.h"
 #include "cs_par_orc_semantics.h"
-
 #include "interlocks.h"
+#include "find_opcode.h"
 
-OENTRY* find_opcode(CSOUND *, char *);
 /***********************************************************************
  * static function prototypes
  */
@@ -61,44 +59,6 @@ static INSTR_SEMANTICS *instr_semantics_alloc(CSOUND *csound, char *name)
 
     return instr;
 }
-
-/***********************************************************************
- * parse time support
- */
-
-//static INSTR_SEMANTICS *curr;
-//static INSTR_SEMANTICS *root;
-
-/* void csp_orc_sa_cleanup(CSOUND *csound) */
-/* { */
-/*     INSTR_SEMANTICS *current = csound-> instRoot, *h = NULL; */
-/*     csp_orc_sa_print_list(csound); */
-/*     while (current != NULL) { */
-
-/*       csp_set_dealloc(csound, &(current->read)); */
-/*       csp_set_dealloc(csound, &(current->write)); */
-/*       csp_set_dealloc(csound, &(current->read_write)); */
-
-/*       h = current; */
-/*       current = current->next; */
-/*       csound->Free(csound, h); */
-/*     } */
-/*     current = csound->instCurr; */
-/*     while (current != NULL) { */
-
-/*       csp_set_dealloc(csound, &(current->read)); */
-/*       csp_set_dealloc(csound, &(current->write)); */
-/*       csp_set_dealloc(csound, &(current->read_write)); */
-
-/*       h = current; */
-/*       current = current->next; */
-/*       csound->Free(csound, h); */
-/*     } */
-
-/*     csound->instCurr = NULL; */
-/*     csound->instRoot = NULL; */
-/* } */
-
 
 static void sanitise_set(CSOUND *csound, struct set_t* p)
 {
