+++ H/inst_ops.h
@@ -26,6 +26,7 @@
 
 #include "csoundCore.h"
 #include "aops.h"
+#include "schedule.h"
 
 typedef struct {
     OPDS    h;
@@ -93,6 +94,7 @@ typedef struct {
   MYFLT *val;
 } PARM_INSTR;
 
+int32_t play_instr(CSOUND *csound, LINEVENT2 *p);
 int32_t kill_instancek(CSOUND *csound, KILLOP *p);
 int32 sa_early(CSOUND *csound, AOP *p);
 int32 sa_offset(CSOUND *csound, AOP *p);
@@ -104,13 +106,5 @@ int32_t pause_instance_opcode(CSOUND *csound, PAUSE_INSTR *p);
 int32_t set_instance_parameter(CSOUND *csound, PARM_INSTR *p);
 int32_t get_instance(CSOUND *csound, DEL_INSTR *p);
 int32_t splice_instance(CSOUND *csound, SPLICE_INSTR *p);
-
-INSDS *instance(CSOUND *, int32_t);
-INSDS *create_instance(CSOUND *csound, int32_t insno);
-void free_instance(CSOUND *csound, INSDS *ip);
-int32_t instr_num(CSOUND *csound, INSTRTXT *instr);
-void free_instr_var_memory(CSOUND* csound, INSDS* ip);
-int32_t init_instance(CSOUND *csound, INSDS *ip, EVTBLK *newevtp);
-int32_t instr_context_check(CSOUND *csound, INSDS *ip, INSDS *insdshead);
 
 #endif
